refactor: extract Heading component from typography to separate file
Move Heading component to dedicated file and enhance with copy
functionality using CopyButton for better user experience when
sharing section links
